Add config validation function
