+++ package.json
@@ -15,6 +15,7 @@
 		"url": "https://www.neodymium.dev"
 	},
 	"bin": "bin.js",
+	"main": "index.js",
 	"types": "types/index.d.ts",
 	"files": [
 		"types"
